LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-spacing/tentative/parsing/text-spacing-trim-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing text-spacing-trim with valid values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('text-spacing-trim', 'auto');
test_valid_value('text-spacing-trim', 'space-all');
</script>
</body>
</html>
